Merge from latest antlr

+++ runtime/Cpp/runtime/src/misc/InterpreterDataReader.cpp
@@ -3,6 +3,10 @@
  * can be found in the LICENSE.txt file in the project root.
  */
 
+#include <sstream>
+#include <cassert>
+#include <vector>
+#include <string>
 #include "atn/ATN.h"
 #include "atn/ATNDeserializer.h"
 #include "Vocabulary.h"
@@ -101,7 +105,7 @@ InterpreterData InterpreterDataReader::parseFile(std::string const& fileName) {
     };
   }
 
-  std::vector<uint16_t> serializedATN;
+  std::vector<int32_t> serializedATN;
 
   std::getline(input, line, '\n');
   assert(line == "atn:");
@@ -115,7 +119,7 @@ InterpreterData InterpreterDataReader::parseFile(std::string const& fileName) {
       number = std::strtoul(&value[1], nullptr, 10);
     else
       number = std::strtoul(value.c_str(), nullptr, 10);
-    serializedATN.push_back(static_cast<uint16_t>(number));
+    serializedATN.push_back(static_cast<int32_t>(number));
   }
 
   ATNDeserializer deserializer;
